+++ src/app/(dashboard)/(chain)/[chain_id]/components/server/primary-info-item.tsx
