+++ src/app/(dashboard)/(chain)/[chain_id]/rpc-edge/components/client/live-stats.tsx
@@ -1,11 +1,62 @@
 "use client";
 
 import { Skeleton } from "@/components/ui/skeleton";
-import { PrimaryInfoItem } from "./PrimaryInfoItem";
 import { CircleCheck, XIcon } from "lucide-react";
 import { ToolTipLabel } from "@/components/ui/tooltip";
 import { CopyButton } from "@/components/ui/CopyButton";
-import { useChainStatswithRPC } from "./useChainStats";
+import { PrimaryInfoItem } from "../../../components/server/primary-info-item";
+import { useState } from "react";
+import { thirdwebClient } from "lib/thirdweb-client";
+import { hostnameEndsWith } from "utils/url";
+import { isProd } from "constants/rpc";
+import { useQuery } from "@tanstack/react-query";
+
+function useChainStatswithRPC(_rpcUrl: string) {
+  const [shouldRefetch, setShouldRefetch] = useState(true);
+
+  let rpcUrl = _rpcUrl.replace(
+    // eslint-disable-next-line no-template-curly-in-string
+    "${THIRDWEB_API_KEY}",
+    thirdwebClient.clientId,
+  );
+
+  // based on the environment hit dev or production
+  if (hostnameEndsWith(rpcUrl, "rpc.thirdweb.com")) {
+    if (!isProd) {
+      rpcUrl = rpcUrl.replace("rpc.thirdweb.com", "rpc.thirdweb-dev.com");
+    }
+  }
+
+  return useQuery({
+    queryKey: ["chain-stats", { rpcUrl }],
+    queryFn: async () => {
+      const startTimeStamp = performance.now();
+      const res = await fetch(rpcUrl, {
+        method: "POST",
+        body: JSON.stringify({
+          jsonrpc: "2.0",
+          method: "eth_blockNumber",
+          params: [],
+          id: 1,
+        }),
+      });
+
+      const json = await res.json();
+      const latency = (performance.now() - startTimeStamp).toFixed(0);
+
+      return {
+        latency,
+        blockNumber: parseInt(json.result, 16),
+      };
+    },
+    refetchInterval: shouldRefetch ? 5 * 1000 : undefined,
+    enabled: !!rpcUrl,
+    refetchOnWindowFocus: false,
+    onError: () => {
+      setShouldRefetch(false);
+    },
+  });
+}
 
 export function ChainLiveStats(props: { rpc: string }) {
   const stats = useChainStatswithRPC(props.rpc);
