+++ src/app/(dashboard)/(chain)/[chain_id]/rpc-edge/page.tsx
@@ -1,7 +1,8 @@
 import { redirect } from "next/navigation";
-import { InfoCard } from "../InfoCard";
-import { ChainLiveStats } from "../overview/components/ChainLiveStats";
+
 import { getChain } from "../utils";
+import { InfoCard } from "../components/server/info-card";
+import { ChainLiveStats } from "./components/client/live-stats";
 
 export default async function Page(props: { params: { chain_id: string } }) {
   const chain = await getChain(props.params.chain_id);
